fix(ios): preserve state on failed polls and correct playlist controls
- refreshCanvasBattery keeps the last known battery/playlist values on
  transport errors and exposes isServerReachable/isRefreshingStatus
  instead of blanking the whole UI after a single failed LAN poll
- error alert resets errorText on dismissal so an identical error can
  be shown again
- polled interval values no longer echo back as UpdatePlaylistInterval
  mutations (track the server-confirmed value)
- quiet-hours toggle and timezone changes now push the new
  updateQuietHours mutation to a running playlist (debounced)
- start() surfaces wokeUp=false as a status instead of failing
- destructive upload (delete-all then upload) asks for confirmation
- one completion haptic per batch instead of one per photo
- fix integer-division scale in ImageProcessor center crop
- drop dead code: widget background-screenshot subsystem, unused
  getBatteryReport, unused PreviewAppViewModel, PlaylistStatus 'stop'
  (replaced by a client-only 'unknown' decode fallback)

+++ ios/Canvas/Canvas/ImageProcessor.swift
@@ -114,12 +114,12 @@ enum ImageProcessor {
         }
         
         // Convertir en coordonnées de pixel
-        let scale = cgImage.width / Int(sourceSize.width)
+        let scale = CGFloat(cgImage.width) / sourceSize.width
         let pixelCropRect = CGRect(
-            x: cropX * CGFloat(scale),
+            x: cropX * scale,
             y: 0,
-            width: croppedWidth * CGFloat(scale),
-            height: sourceSize.height * CGFloat(scale)
+            width: croppedWidth * scale,
+            height: sourceSize.height * scale
         ).integral
         
         guard let croppedCGImage = cgImage.cropping(to: pixelCropRect) else {
